+++ xc3_wgpu_batch/src/main.rs
@@ -315,6 +315,7 @@ fn calculate_camera_data(width: u32, height: u32, translation: Vec3, rotation: V
 
     CameraData {
         view,
+        projection,
         view_projection,
         position,
     }